move opcodes to separate file
